+++ package/utils/requireOrError.js
@@ -7,7 +7,7 @@ const requireOrError = (moduleName) => {
     return require(moduleName)
   } catch (error) {
     throw new Error(
-      `[SHAKAPACKER]: ${moduleName} is required for ${config.bundler} but is not installed. View Shakapacker's documented dependencies at https://github.com/shakacode/shakapacker/tree/main/docs/peer-dependencies.md`
+      `[SHAKAPACKER]: ${moduleName} is required for ${config.assets_bundler} but is not installed. View Shakapacker's documented dependencies at https://github.com/shakacode/shakapacker/tree/main/docs/peer-dependencies.md`
     )
   }
 }